+++ example/lib/src/module/text_page.dart
@@ -12,39 +12,59 @@ class TextPage extends StatelessWidget {
           appBar: AppBarText('Text'),
           padding: const EdgeInsets.fromLTRB(20, 0, 20, 20),
           children: [
+            const Partition('BText  DefaultTextStyle'),
+            DefaultTextStyle(
+                style: context.textTheme.bodyLarge ?? const TextStyle(),
+                child: const Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceAround,
+                    children: [
+                      BText('BText'),
+                      BText.rich(
+                          text: 'Start text  ',
+                          texts: ['BText', ' * ', 'rich']),
+                    ])),
             const Partition('BText'),
             const BText('BText'),
             const Partition('BText.rich'),
             const BText.rich(
-                style: TextStyle(color: Color(0xFF42A5F5)),
+                style: TextStyle(color: Colors.green),
+                color: Colors.red,
                 texts: [
                   'BText',
                   ' * ',
-                  'rich',
+                  'rich'
                 ],
                 styles: [
-                  TextStyle(color: Color(0xFFD32F2F)),
+                  TextStyle(color: Colors.purpleAccent),
                   TextStyle(color: Colors.blue),
-                  TextStyle(color: Color(0xFFFFC400)),
+                  TextStyle(color: Colors.yellow),
                 ]),
+            const BText.rich(
+                style: TextStyle(color: Colors.red),
+                color: Colors.green,
+                texts: ['BText', ' * ', 'rich'],
+                styles: [TextStyle(color: Colors.purpleAccent)]),
+            const BText.rich(
+                style: TextStyle(color: Colors.red),
+                color: Colors.green,
+                texts: ['BText', ' * ', 'rich']),
             const Partition('RText'),
-            RText(
-                style: const TextStyle(color: Color(0xFF42A5F5)),
-                texts: const [
-                  'RText',
-                  ' = ',
-                  'RichText魔改版',
-                ],
-                styles: const [
-                  TextStyle(color: Color(0xFFD32F2F)),
-                  TextStyle(color: Colors.white),
-                  // TextStyle(color: Color(0xFFFFC400)),
-                ]),
+            RText(style: const TextStyle(color: Colors.green), texts: const [
+              'RText',
+              ' = ',
+              'RichText魔改版',
+            ], styles: const [
+              TextStyle(color: Colors.purpleAccent),
+              TextStyle(color: Colors.blue),
+              TextStyle(color: Colors.yellow),
+            ]),
             const Partition('BText set color'),
             const BText('BText', color: Colors.red),
             const Partition('BText set style color'),
             const BText('BText',
-                color: Colors.red, style: TextStyle(color: Colors.blue)),
+                color: Colors.red,
+                style: TextStyle(color: Colors.blue),
+                useStyleFirst: true),
             const Partition('BText set style color useStyleFirst=false'),
             const BText('BText',
                 useStyleFirst: false,